fix CI

+++ sum/contracts/assembly/contracts/run.ts
@@ -30,8 +30,8 @@ export function constructor(sumAddress: StaticArray<u8>): void {
  */
 export function main(_: StaticArray<u8>): void {
   const address = new Address(Storage.get('address'));
-  const a: i32 = 2;
-  const b: i32 = 3;
+  const a: i64 = 2;
+  const b: i64 = 3;
   const values = new Args().add(a).add(b);
   call(address, 'sum', values, 5000000);
   return;
